fix: refactor - move information card code to new file
